Added a HashSet<Token> instance variable to the Player class which will be used to keep track of the tokens that the player has collected.
Added accessor and mutator for the HashSet<Token> instance variable in
Player Class.

Edited the movePlayer() method in Class Board so that it checks if the
tile the player moved to contains a token, if so it will check if the
token is face down and turn it over if so, and then will check the
number of the Token and compare it with the current number that the
players are looking for. If the numbers match then the token will be
added to the player's HashSet<Token> and removed from the tile. The
current number will then be incremented. 

Moved initializeTokens() from initializeBoard() to Board constructor.

Other misc changes